Change the behavior of the interface of Textrepo::Repository
Target API: Textrepo::Repository#update(timestamp, text)

Repository class is an abstract class that is not intended to be
instantiated.  So, this class doesn't change in this commit.
Instead, the concrete class Textrepo::FileSystemRepository was
changed.

Before: update(timestamp, text) updates only the content of the
file in the repository.

After:  update(timestamp, text) updates not only the content,
but also the associated timestamp to the current time.