[impl] spec change for Timestamp
Date resolution for a timestamp attached to a file is different
depends on a file system.  On macOS, HFS+ adopts 1 sec and APFS adopts
1 nano sec for the resolution.

So, if textrepo uses millisecond part in its timestamp, it may cause a
crash on some file systems.

The new spec for the Timestamp is the following:

- the resolution is 1 sec
- use "suffix" to import several files those have the same birthtime
  (in the 1 sec resolution)
- "suffix" is used as a serial number, and it converts into a string
  such "012"
- "suffix" is put the tail of the pathname like "202010101802_012"
- the range of "suffix" is from "001" to "999"

And the most important spec is:

- a suffix is optional

Though Timestamp accepts the order to use a suffix, it does not try to
add it automatically.  The upper layer to use Textrepo must take care
how to use it.  The only one rule of Textrepo to store text is that
each text has a unique timestamp in the repository.

Since a suffix is optional, there may be 2 types of timestamp string,
that is one has a suffix part and the other has no suffix.

- the comparison of 2 timestamp is done in the rule such as using
  suffix as fraction of sec:

      - "20201122010203" < "20201122010203_012"
      - "20201122010203_001" < "20201122010203_999"
      - "20201122010203_999" < "20201122010204"

All the modifications have been applied to Timestamp class.  And tests
were also updated to verify the new spec described above.